nle/agent/requirements.txt
nle
hydra-core 
hydra_colorlog
wandb
einops
torch
numpy
